Change Android ADM JNI method registration management
This commit aligns the Android ADM JNI method registration with the
class registration and unregistration. This change allows users
to create an Android ADM immediately after a previous one was destroyed.
Previously, there were cases where the destruction of one ADM
would unregister JNI methods needed in a freshly created instance and
result in a crash.

Bug: webrtc:7587
Change-Id: I54640d30a7ec22c156e06a11fa231b236a5f16be