Update patch set 1
Patch Set 1:

I validated the existing instrumentation tests pass as expected. While the steps to reproduce this bug are not clear, if necessary, I can add a test that creates and destroys multiple Android ADMs to validate the behavior does not result in a crash.

Patch-set: 1
Reviewer: Gerrit User 5112 <5112@58829da1-049c-39fb-b951-ebdcd0984767>
Work-in-progress: false