Update patch set 1
Patch Set 1: Code-Review+1

(5 comments)

Looks good to me.
A question on whether this is a generic issue we should DCHECK on, a suggestion on simplifying the tests, and a few comments on English grammar.

Given that this is a SEGV, I'm giving it the +1 and letting you judge whether the generic suggestion needs to be taken on in a later CL.

Patch-set: 1
Reviewer: Gerrit User 5634 <5634@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Code-Review=+1