Update patch set 1
Patch Set 1:

I am not an owner in rtc_base and I am not sure if we really want to (or can) add this functionality to WebRTC at this stage. The bug does not contain a final decision from any WebRTC member if the support should be added or not.

anatolid@: Can you drive this?

Patch-set: 1
CC: Gerrit User 5526 <5526@58829da1-049c-39fb-b951-ebdcd0984767>