Update patch set 1
Patch Set 1:

I don't believe it to be appropiate to do this in WebRTC unless CSRC 42 has a special meaning in the WebRTC standard itself.

If CSRC 42 is application-specific, then the application should be able to compensate for it on its own. E.g. the caller of `getContributingSources()` in JavaScript doing the merging rather than doing it here.

Patch-set: 1
Reviewer: Gerrit User 8392 <8392@58829da1-049c-39fb-b951-ebdcd0984767>