Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > This seems like a fine solution but obviously we need to have internal bot pass as well.
> > 
> Thank you, that what I was looking for. Since you ok with general approach I'll work on the internal bot.
> 
> > Where possible we should probably also just directly use the generated enum instead of wrapping it in another enum.
> I've sketched that approach too: https://webrtc-review.googlesource.com/c/src/+/167301
> there is no generate enum though, just a bunch of int values packed together.
> Will go with that approach if you prefer it. (for sure still need to add same support for it in the internal bot).

I think it is preferable at least where the enum is not part of any exposed API surface. As long as any tests are not broken by the removal of the enum I think we should be good to go with that one, I left a few comments in the CL.

> Can you please comment on one more detail: chromium script generate java that adds dependency on androidx annotation. Do you think that can be a problem?

I don't think it is a problem (as long all the tests pass). If not, let me know, we might have to add a dependency somewhere.

Patch-set: 2
Reviewer: Gerrit User 5112 <5112@58829da1-049c-39fb-b951-ebdcd0984767>