Update patch set 2
Patch Set 2:

I'm not wery well versed with the details of how Chromium interacts with WebRTC. But it seems to me that if we want to expose it, that we should also move it to the api folder and do the associated level of cleanup?

In general I'd like to see less dependencies on WebRTC code in Chromium since it creates large maintenance overhead (requires 2-3 extra CLs split over two repos for even simple changes, making a change take roughly a week rather than a day). In this case it might still make sense, or maybe it would make sense to look at why it's needed and see if we could be more explicit about the time handling in the components that require it, removing the need to fiddle with a global clock instance.

Ideally I'd prefer if higher level test abstractions were used, such as the peer scenario framework, that way we hide all details but those that strictly required (the injection of the blink-specific audio sink) and ensure that any changes in WebRTC that would break the integration also will trigger a test failure (currently such breakages could be missed in the proposed unit test).

Patch-set: 2
Reviewer: Gerrit User 5531 <5531@58829da1-049c-39fb-b951-ebdcd0984767>