Update patch set 4
Patch Set 4:

Replying to your comment before reviewing the CL.

Also speaking of manually verifying stats, did you try importing this CL into chromium and checking chrome://webrtc-internals stats page? That may give a nice overview.

> Patch Set 4:
> 
> @hbos, tests are fixed now and I also went through all the stats fields in RTCOutboundRTPStreamStats and here is the situation after applying the CL:
> 
> RTCStatsMember<uint32_t> ssrc;                                  //OK
> RTCStatsMember<bool> is_remote;                                 //OK, hardcoded to false, https://bugs.webrtc.org/7066

isRemote doesn't exist in the spec anymore and is not needed (remote RTP dictionaries 

> RTCStatsMember<std::string> media_type;                         //OK
> RTCStatsMember<std::string> kind;                               //OK
> RTCStatsMember<std::string> track_id;                           //OK
> RTCStatsMember<std::string> transport_id;                       //OK
> RTCStatsMember<std::string> codec_id;                           //OK, not in spec?

It's in the spec, but it's in a parent dictionary in the hierarchy (RTCRtpStreamStats).
So if this was added to RTCOutboundRtpStreamStats as well then it would appear twice, which would be a bug.

> RTCStatsMember<uint32_t> fir_count;                             //OK
> RTCStatsMember<uint32_t> pli_count;                             //OK
> RTCStatsMember<uint32_t> nack_count;                            //OK
> RTCStatsMember<uint32_t> sli_count;                             //Not wired. https://bugs.webrtc.org/7066
> RTCStatsMember<uint64_t> qp_sum;                                //OK

Are these counted on a per-layer basis?

>   
> RTCStatsMember<std::string> media_source_id;                    //OK
> RTCStatsMember<std::string> remote_id;                          //OK

The remote ID is the RTCP feedback for this stream.
If we have per-layer outbound-rtp then I guess we would have per-layer remote-inbound-rtp objects referenced by this field. Do we do this correctly, or do we have a single remote-inbound-rtp for all layers?

> RTCStatsMember<std::string> rid;                                //OK, new
> RTCStatsMember<uint32_t> packets_sent;                          //OK, but numbers look suspicious for rtx

Suspicious how?
This TODO sounds... suspicious: https://source.chromium.org/chromium/chromium/src/+/master:third_party/webrtc/media/engine/webrtc_video_engine.cc;l=2417;drc=b892cf58e162a8f66cd76d7472f129fe0fb6a7d1;bpv=1;bpt=1?originalUrl=https:%2F%2Fcs.chromium.org%2F

> RTCStatsMember<uint64_t> retransmitted_packets_sent;            //OK
> RTCStatsMember<uint64_t> bytes_sent;                            //OK
> RTCStatsMember<uint64_t> header_bytes_sent;                     //OK, but numbers look suspicious for rtx

Same question as above - suspicious how?

> RTCStatsMember<uint64_t> retransmitted_bytes_sent;              //OK
> RTCStatsMember<double> target_bitrate;                          //Not wired. https://bugs.webrtc.org/7066
> RTCStatsMember<uint32_t> frames_encoded;                        //OK
> RTCStatsMember<uint32_t> key_frames_encoded;                    //OK
> RTCStatsMember<double> total_encode_time;                       //Same for all. TODO: OnEncodedFrameTimeMeasured doesn't have ssrc
> RTCStatsMember<uint64_t> total_encoded_bytes_target;            //Accumulated. TODO: OnSendEncodedImage has ssrc



> RTCStatsMember<uint32_t> frame_width;                           //OK, new
> RTCStatsMember<uint32_t> frame_height;                          //OK, new
> RTCStatsMember<double> frames_per_second;                       //OK, new
> RTCStatsMember<double> total_packet_send_delay;                 //OK

These are correctly reported for each layer?

> RTCStatsMember<std::string> quality_limitation_reason;          //Same for all but OK?
> RTCStatsMember<uint32_t> quality_limitation_resolution_changes; //Same for all but OK?
> RTCStatsMember<std::string> content_type;                       //Same for all but OK?

Yes, that makes sense. The current implementation adapts the source, so the effect is visible on all layers.

There is also simulcast logic about enabling or disabling layers that can kick in if you are bandwidth limited, which should count as a quality limitation.
For that edge case I would file a bug, assign ilnik@webrtc.org to it, and reference it here.
The bug would be that in some edge cases we may not be reporting adaptation even though we are adapting, but I think that is an existing bug.

> RTCStatsMember<std::string> encoder_implementation;             //Same for all but OK?

That's also OK. I think there is some recent work that may try to use SW encoding on one of the layers in some edge case. I would file a bug and assign eshr@google.com and reference it. In either case it would be an edge case that is an existing bug.

Patch-set: 4
Reviewer: Gerrit User 5142 <5142@58829da1-049c-39fb-b951-ebdcd0984767>