Update patch set 4
Patch Set 4:

(3 comments)

> Patch Set 4:
> 
> > > RTCStatsMember<uint64_t> qp_sum;                                //OK
> > 
> > Are these counted on a per-layer basis?
> At least the wiring supports it now as layer based, but haven't looked deeper than send_statistcs_proxy.cc
> 
> for rtx
> > 
> > Suspicious how?
> 
> Looked at chrome://webrtc-internals while running the simulcast example provided in the bug and the numbers were constantly growing while I don't think rtx was actually used to retransmit anything.

OK. The playground is doing some funny SDP tricks to set up simulcast without the involvement of a middlebox, so it may very well get confused and send RTX packets. Let's expose the metrics. Either they will be useful for debugging actual RTX or we can worry about invalid RTX counters in the future, let's assume they are good for now.

> 
> > > RTCStatsMember<uint32_t> frame_width;                           //OK, new
> > > RTCStatsMember<uint32_t> frame_height;                          //OK, new
> > > RTCStatsMember<double> frames_per_second;                       //OK, new
> > > RTCStatsMember<double> total_packet_send_delay;                 //OK
> > 
> > These are correctly reported for each layer?
> 
> The 3 new ones I added seemed to work correctly, even though for rtx they were all 0, which is probably expected in case no rtx frames were sent.

Is the RTX stream spawning an additional "outbound-rtp" object? It makes sense to have metrics for the RTX stream, but I'm not really sure what what the intent of the stats spec is if this should be another outbound-rtp or if this case was not considered. Can we we skip the RTX stream for now?

Patch-set: 4
Reviewer: Gerrit User 5142 <5142@58829da1-049c-39fb-b951-ebdcd0984767>