Update patch set 4
Patch Set 4:

The fact that RTX was baked into the same outbound-rtp stream and now won't further increases the risk of breakage when we land this CL even for non-simulcast cases. I wonder if that can be avoided somehow or if we just have to be careful about PSAs. +hta@ to also take a look.

Patch-set: 4
Reviewer: Gerrit User 5634 <5634@58829da1-049c-39fb-b951-ebdcd0984767>
Reviewer: Gerrit User 5142 <5142@58829da1-049c-39fb-b951-ebdcd0984767>