Update patch set 10
Patch Set 10:

(4 comments)

Because of the complexity of this CL and the risks associated with shipping simulcast stats I would like this CL to have a flag deciding whether or not to enable per-layer infos.

I think this helps making progress. This allows us to:

- Land this code and before shipping simulcast stats, getting the GetAggregatedVideoSenderInfo() code exercised in cases where we would not expect any regressions. This is harder to do if we ship simulcast stats at the same time because then we have to figure out if a change was intentional or a regression.

- When we do ship simulcast stats, if for some reason we have to revert it, we can revert a smaller "unflagging CL" instead of reverting a large and complex CL.

- When we do ship simulcast stats, possibly add a few more tests that can be reviewed separately from this CL.

Note that I am still working on https://webrtc-review.googlesource.com/c/src/+/170826 which need to land first, but it is on its way.

Patch-set: 10
Reviewer: Gerrit User 5142 <5142@58829da1-049c-39fb-b951-ebdcd0984767>