Update patch set 3
Patch Set 3:

(1 comment)

> So, this resolution_bitrate_limits_ isn't used anywhere except for GetEncoderInfo right now?

That is correct, but GetEncoderInfo().resolution_bitrate_limits is used by EncoderInfo.GetEncoderBitrateLimitsForResolution(), whose output is used here:
https://source.chromium.org/chromium/_/webrtc/src.git/+/005d0fff3b9c56477bb591855fd62639bd8c9eca:video/video_stream_encoder.cc;l=1625
So in effect, this CL is actually the final wire-up that allows external users of LibvpxVp8Encoder to affect what resolution/bitrate limits are used in the VideoStreamEncoder.

> 
> If the CL wiring it up isn't that far off, maybe you should add the stuff you want to do with this new
> setting in this CL. Otherwise it's not great to have an essentially unused setting in the code. WDYT?

I agree with this methodology, but it's not applicable for this particular CL as the wire-up is already checked in (see above).

Patch-set: 3
Reviewer: Gerrit User 5023 <5023@58829da1-049c-39fb-b951-ebdcd0984767>