Update patch set 24
Patch Set 24:

(10 comments)

Thanks for the comments!
I'm addressed them in patchsets.

I also have reverted part of the changes that ensure that audioproc_f did work even with this build flag. The reason for the revert was that it broke the API. With the current change it can run but will fail on a CHECK since the APM never gets created.

There is still one test bot failing due to the fuzzers needing access to test-only code (which turned out not to be built for the fuzzers).

Patch-set: 24
Reviewer: Gerrit User 5125 <5125@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Commit-Queue=0