Fixes padding handling while only sending non-paced audio.
There were two issues in PacingController that triggered when only
non-paced audio was sent while pacing was requestest and there was an
existing padding debt.

One issue was NextSendTime(), which should trigger an immediate send if
audio is not to be paced and the next packet in the queue is audio.
Previously it only did this while congested.

Second, EnqueuePackets() would reset the last process time if the queue
and media budgets were empty. This was done without reducing the
padding debt.

Bug: webrtc:10809
Change-Id: I116169522c215257febd32e17abab45f1a7d609f