Update patch set 15
Patch Set 15:

> > Because of the complexity of this CL and the risks associated with shipping simulcast stats I would like this CL to have a flag deciding whether or not to enable per-layer infos.
> 
> Adding the flag might be easy, but flipping it would be still pain for tests. I guess the tests should also have two paths then depending on the flag.

In some of the tests I had to check the flag to make it pass for both settings, but didn't add any TEST_P to cover both settings at the same time. In my opinion that would be overkill given it is a temporary flag and in normal use it is not meant to be flipped at runtime. Also chromium side tests still need modifications to support flipping of the flag. For now I verified that in both settings all webrtc tests pass, but of course the question is how long we need to keep that flag. During that time there is a chance that some test may get broken in 'enable_simulcast_stats == true' setting.

Patch-set: 15
Reviewer: Gerrit User 9279 <9279@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Commit-Queue=0