Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review+1
> 
> This change looks good, however I'm curious how you ever could run into this scenario. As far as I can tell, the only way a SocketDispatcher would be reused in this way is if you call Connect on it twice; is that what's happening? And if so, where?

Yes, it happens in the peerconnection example. I'm not sure if it happens anywhere in the main codebase.

In peer_connection_client.cc, the socket hanging_get_ is closed and connected several times. It does this in the same socket's read callback (OnHangingGetRead), which is where we hit this bug

Patch-set: 2
CC: Gerrit User 11299 <11299@58829da1-049c-39fb-b951-ebdcd0984767>