Reduces unnecessary thread wakeups in TaskQueuePacedSender.
This CL is functionally a noop but may reduce thread wakupes in some
cases.

In particular, consider a send task scheduled for time T. While waiting
for that, a higher-priority packet than the top of the current queue is
added (e.g. and audio packet), and a send is executed immediately.

After sending it resets the field indicating that a scheduled task is
expected at time T. It then polls NextSendTime() and schedules a new
task, likely at or very close to T. Causing unnecessary task queue
churn and behavior that is more difficult to reason about.

Bug: webrtc:10809
Change-Id: Ic5706f2cc06df3f27cc3e7b473d4de29a669473b