Adds support for generating FEC at the RTP egress stage.
By delaying FEC generation until protected packets are actually sent,
we can fully protect header extensions populated at the egress stage,
such as transport wide sequence number, abs send time, video timing,
etc.

The receive side needs to know that this is how sender is behaving,
as otherwise it needs to zero the extensions in the recover mechanism
in order to prevent corruptions. Therefor, I added a header extension
that can be negotiated. It's not actually transmitted, its presence
just means signals that header should be protected.

This CL works, but it's already too large and still needs cleanup and
more test coverage, so going to keep it for context only and will split
it into several smaller parts.

Change-Id: I9040e5f40ceac9010962ebb9fefac4ecaddd8018