Update patch set 6
Patch Set 6:

> Patch Set 6: Code-Review+1
> 
> Cool! Me and srte@ actually discussed this a while back to avoid reconfigurations. IIRC, absolute_capture_time extension is supposed to be sent once every second, maybe we should exclude that from the overhead? Are there any other extensions like that?
> 
> Changes in audio_send_stream look good. Sorry for the delay :)

Ah, I didn't notice that that extension isn't always intended to be sent. Is that already so or something in the works?
There are several things on the video packets that are not always sent, in fact some extensions are only used on the last packet of a video frame. That's why I added the rolling averaging, but if _any_ change causes audio reconfiguration that risks triggering a reconfig on every rtcp message at the moment.
Thinking we should either have a one second period timer to update, or for audio just look at static max overhead where abs capture time is either always or never included.
+srte to cc in case there's any comment on the aforementioned discussion

Patch-set: 6
CC: Gerrit User 5531 <5531@58829da1-049c-39fb-b951-ebdcd0984767>