Update patch set 7
Patch Set 7:

Confirmed that with transport seq used, the rtcp messages can be quite frequent and due to the sliding averaging this can trigger a reconfig every time.
Added a 1s max update frequency in that path, as this matches abs capture time. Even if they don't sync up perfectly this still results in ~1 update per second instead of two as it is now. If no dynamically updated extension is used, the average won't change either.
ptal

Patch-set: 7