Reland remaining changes from https://webrtc-review.googlesource.com/c/src/+/172847:
This applies the changes from the above CL to the forked files and
switches call.cc over to using that implementation.

This CL includes all of the DCHECKs except for ones that failed on iOS
where decoded frames might arrive on a callback thread outside of webrtc.
Those DCHECKs are disabled for now and I added comments to make it clear
where this happens.

Bug: webrtc:11489
Change-Id: I93ad560500806ddd0e6df1448b1bcf5a1aae7583