Update patch set 1
Patch Set 1:

Hey Julien,

This is a small follow up to the X11 screen share patch. Users with 3 or more monitors are having crashes when selecting monitors. I noticed that even without crashing on my machine with three monitors the frame in the selection dialog is not sized correctly when the third monitor has a different resolution from the others. This fixes that for me.

Thought I would just pull you in first since you were the most active reviewer on this last time.
Thank you.

Patch-set: 1
Reviewer: Gerrit User 6236 <6236@58829da1-049c-39fb-b951-ebdcd0984767>
Work-in-progress: false