Update patch set 3
Patch Set 3:

(2 comments)

> Patch Set 1:
> 
> (1 comment)
> 
> > Patch Set 1:
> > 
> > Hey Julien,
> > 
> > This is a small follow up to the X11 screen share patch. Users with 3 or more monitors are having crashes when selecting monitors. I noticed that even without crashing on my machine with three monitors the frame in the selection dialog is not sized correctly when the third monitor has a different resolution from the others. This fixes that for me.
> > 
> > Thought I would just pull you in first since you were the most active reviewer on this last time.
> > Thank you.
> 
> Hi Trevor,
> Great that you found some times, I am happy to review, see my comment below.
> Thx!

Thanks for the review Julien. I have added more detail to the description, and added a comment to the code. Also there was an existing comment I found confusing when working on this. I'm not sure if this would be a good time to revise it. I have an extra note about that below.

Thank you.

Patch-set: 3